print format fixes, gpu backbuffer, process scheduling index stability fixes

+++ shared/process.h
@@ -20,7 +20,7 @@ typedef struct {
     uint64_t sp;
     uint64_t pc;
     uint64_t spsr; 
-    uint64_t id;
+    uint16_t id;
     enum process_state { STOPPED, READY, RUNNING, BLOCKED } state;
     input_buffer_t input_buffer;
 } process_t;
